Modular map (#23)
* spawn track

track is spawned and destroyed. now must work on setting target track to the next track in the gameMode_endless trackList array

* track spawns, camera works

* comments

added comments to worm blueprint event graph

* worm is all slidey now

but in a good way

* all the comments

so pretty so legible (GameMode_endless, chrisCraft)

* fix track spawning

lerping is still weird during transitions, but the track spawns invisible and becomes visible as it gets closer (in the track array). Track overlaps with worm and boat only trigger when the track is the current track

* final save before switching computers